+++ src/converters/pip/mod.rs
@@ -1,11 +1,10 @@
 mod dependencies;
 
-use crate::converters::pip::dependencies::get;
 use crate::converters::pyproject_updater::PyprojectUpdater;
 use crate::converters::Converter;
 use crate::converters::ConverterOptions;
 use crate::schema::pep_621::Project;
-use crate::schema::pyproject::DependencyGroupSpecification;
+use crate::schema::pyproject::{DependencyGroupSpecification, PyProject};
 use crate::schema::uv::Uv;
 use crate::toml::PyprojectPrettyFormatter;
 use indexmap::IndexMap;
@@ -26,6 +25,10 @@ pub struct Pip {
 
 impl Converter for Pip {
     fn build_uv_pyproject(&self) -> String {
+        let pyproject_toml_content =
+            fs::read_to_string(self.get_project_path().join("pyproject.toml")).unwrap_or_default();
+        let pyproject: PyProject = toml::from_str(pyproject_toml_content.as_str()).unwrap();
+
         let dev_dependencies = dependencies::get(
             &self.get_project_path(),
             self.dev_requirements_files.clone(),
@@ -73,7 +76,7 @@ impl Converter for Pip {
             pyproject: &mut updated_pyproject,
         };
 
-        pyproject_updater.insert_pep_621(&project);
+        pyproject_updater.insert_pep_621(&self.build_project(pyproject.project, project));
         pyproject_updater.insert_dependency_groups(dependency_groups.as_ref());
         pyproject_updater.insert_uv(&uv);
 
@@ -126,7 +129,7 @@ impl Converter for Pip {
             return None;
         }
 
-        if let Some(dependencies) = get(
+        if let Some(dependencies) = dependencies::get(
             self.get_project_path().as_path(),
             self.requirements_files
                 .clone()